Update code, unit test, README
