+++ app/endpoints/user_routes.py
@@ -2,7 +2,7 @@
 from fastapi import FastAPI, HTTPException, APIRouter
 
 # import model
-from app.models.UsedCarModel import UsedCar
-from app.models.UserModel import User
+from app.models.used_car import UsedCar
+from app.models.user import User
 
 router = APIRouter()
