+++ app/server.py
@@ -2,7 +2,6 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-
 # import routers
 from app.endpoints import car_routes, user_routes
 
@@ -12,6 +11,7 @@ app = FastAPI()
 origins = [
     "http://localhost:3000",
 ]
+
 app.add_middleware(  
     CORSMiddleware,
     allow_origins=origins,
