+++ app/services/connect.py
@@ -7,7 +7,6 @@ load_dotenv()
 mongoUser = os.getenv("MONGO_USER")
 mongoPassword = os.getenv("MONGO_PASSWORD")
 
-
 client = motor_asyncio.AsyncIOMotorClient(f'mongodb+srv://{mongoUser}:{mongoPassword}@cluster0.sstu5.mongodb.net/test')
 database = client.CompCar
 car_collection = database.usedCar
